+++ common/src/main/java/smartin/miapi/mixin/EnchantmentMixin.java
@@ -5,7 +5,6 @@ import net.minecraft.enchantment.Enchantment;
 import net.minecraft.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import smartin.miapi.config.MiapiConfig;
 import smartin.miapi.item.modular.ModularItem;
 import smartin.miapi.modules.properties.EnchantmentProperty;
@@ -22,7 +21,7 @@ public class EnchantmentMixin {
         if (stack.getItem() instanceof ModularItem) {
             boolean acceptable = EnchantmentProperty.isAllowed(stack, enchantment);
             if (MiapiConfig.INSTANCE.server.enchants.lenientEnchantments) {
-                acceptable = acceptable || cir.getReturnValue();
+                acceptable = acceptable || original;
             }
             return acceptable;
         }