+++ common/src/main/java/smartin/miapi/mixin/PlayerEntityMixin.java
@@ -4,9 +4,7 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.sound.SoundEvent;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -15,9 +13,6 @@ import smartin.miapi.events.MiapiEvents;
 
 @Mixin(PlayerEntity.class)
 public abstract class PlayerEntityMixin {
-
-    @Shadow
-    public abstract void playSound(SoundEvent sound, float volume, float pitch);
 
     @Inject(
             method = "attack",