+++ common/src/main/java/smartin/miapi/mixin/client/ElytraFeatureRendererAccessor.java
@@ -2,7 +2,6 @@ package smartin.miapi.mixin.client;
 
 import net.minecraft.client.render.entity.feature.ElytraFeatureRenderer;
 import net.minecraft.client.render.entity.model.ElytraEntityModel;
-import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;