+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v1.1.29 (1.20.1)
 - added ability to fake repair ingredient more convincingly for badly written mods
 - improved mixin logic to help with compatiblity
+- fixed elytra not being animated